requirements.txt
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
backoff==2.2.1
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.128.0
ffmpy==1.0.0
filelock==3.20.1
filetype==1.2.0
fsspec==2025.12.0
gradio==6.2.0
gradio_client==2.0.2
greenlet==3.3.0
groovy==0.1.2
h11==0.16.0
haystack-ai==2.21.0
haystack-experimental==0.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.2.3
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lazy_imports==1.2.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
more-itertools==10.8.0
networkx==3.4.2
nltk==3.9.2
numpy==2.2.6
openai==2.14.0
orjson==3.11.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
playwright==1.57.0
posthog==7.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
pyee==13.0.0
Pygments==2.19.2
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.21
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.30.0
safehttpx==0.1.7
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.1
starlette==0.50.0
tenacity==9.1.2
textstat==0.7.12
tomlkit==0.13.3
tqdm==4.67.1
typer==0.21.0
typer-slim==0.21.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
